+++ lib/src/facade/transaction.dart
@@ -36,11 +36,12 @@ final class Transaction {
     DocumentReference<T> ref,
     T data, {
     bool merge = false,
+    List<String>? mergeFields,
     Precondition? precondition,
   }) {
     _writesStarted = true;
     _writes.add(stageSet(ref.path, ref._converter.toMap(data),
-        merge: merge, precondition: precondition));
+        merge: merge, mergeFields: mergeFields, precondition: precondition));
   }
 
   /// Stages an update operation.